refactor: Separate accountParser
