+++ app/gateways/db/postgres/accounts/fetch.go
@@ -3,23 +3,9 @@ package accountspg
 import (
 	"context"
 	"fmt"
-	"time"
 
-	"github.com/google/uuid"
 	"github.com/jpgsaraceni/suricate-bank/app/domain/entities/account"
-	"github.com/jpgsaraceni/suricate-bank/app/vos/cpf"
-	"github.com/jpgsaraceni/suricate-bank/app/vos/hash"
-	"github.com/jpgsaraceni/suricate-bank/app/vos/money"
 )
-
-type queryReturn struct {
-	id        string
-	name      string
-	cpf       string
-	secret    string
-	balance   int
-	createdAt time.Time
-}
 
 func (r Repository) Fetch(ctx context.Context) ([]account.Account, error) {
 	const query = `
@@ -44,14 +30,14 @@ func (r Repository) Fetch(ctx context.Context) ([]account.Account, error) {
 	var accountList []account.Account
 
 	for rows.Next() {
-		var account queryReturn
+		var accountReturned queryReturn
 		err := rows.Scan(
-			&account.id,
-			&account.name,
-			&account.cpf,
-			&account.secret,
-			&account.balance,
-			&account.createdAt,
+			&accountReturned.id,
+			&accountReturned.name,
+			&accountReturned.cpf,
+			&accountReturned.secret,
+			&accountReturned.balance,
+			&accountReturned.createdAt,
 		)
 
 		if err != nil {
@@ -59,37 +45,15 @@ func (r Repository) Fetch(ctx context.Context) ([]account.Account, error) {
 			return nil, fmt.Errorf("%w: %s", errScanningRows, err.Error())
 		}
 
-		parsedAccount, err := account.parse()
+		parsedAccount, err := accountReturned.parse()
 
 		if err != nil {
 
 			return nil, fmt.Errorf("%w: %s", errParse, err.Error())
 		}
 
-		fmt.Println(parsedAccount)
-
 		accountList = append(accountList, parsedAccount)
 	}
 
 	return accountList, nil
-}
-
-func (q *queryReturn) parse() (account.Account, error) {
-	var parsedAccount account.Account
-
-	// id, err := uuid.Parse(q.id)
-
-	// if err != nil {
-
-	// 	return account.Account{}, fmt.Errorf("%w: %s", errIdParse, err.Error())
-	// }
-
-	parsedAccount.Id = account.AccountId(uuid.MustParse(q.id))
-	parsedAccount.Name = q.name
-	parsedAccount.Cpf, _ = cpf.NewCpf(q.cpf)
-	parsedAccount.Balance, _ = money.NewMoney(q.balance)
-	parsedAccount.Secret = hash.Parse(q.secret)
-	parsedAccount.CreatedAt = q.createdAt
-
-	return parsedAccount, nil
 }
